requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Moondream2 Vision Model
transformers==4.37.0
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
einops==0.7.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# For handling images
aiofiles==23.2.1
